+++ styles.css
@@ -8,3 +8,23 @@ table {
   width: 300px;
   color: #000;
 }
+
+.pointer {
+  cursor: pointer;
+}
+
+.identifier {
+  max-width: 50px;
+  display: block-inline;
+  font-size: 50%;
+  padding: 6px;
+  color: #999;
+  border: 1px solid #999;
+  font-family: monospace;
+}
+
+.number {
+  font-family: monospace;
+  font-size: 80%;
+  font-weight: 100;
+}
